Fall back to Android's implementation of getResourcesForApplication
PackageManager.getResourcesForApplication can be used on an Android device
to get the resources for a not-installed apk, but it only works for
installed apks in robolectric.

This falls back to the platform implementation for not-installed apks,
while maintaining the existing functionality for installed apks.

This change only affects the behavior in binary resource mode. In legacy
resource mode, there should be no change in behavior.

PiperOrigin-RevId: 246042251